#458: ManageAccountView 추가
- 회원탈퇴 버튼과 로그아웃 버튼을 담는 뷰 객체를 추가하였습니다.
